Revert "Stats: Add API for stats page module status management (#31057)"
This reverts commit 6f6bc679fc87d6f1e63844c2c68af202d16eb6d1.
