--- projects/packages/stats-admin/changelog/add-stats-settings-support
@@ -1,4 +1,0 @@
-Significance: minor
-Type: added
-
-Stats Admin: add endpoint to save Stats Dashboard settings
