--- projects/packages/stats/changelog/add-stats-settings-support
@@ -1,4 +1,0 @@
-Significance: minor
-Type: added
-
-Stats: add modules to Options
