+++ projects/packages/stats/tests/php/test-options.php
@@ -46,7 +46,6 @@ class Test_Options extends StatsBaseTestCase {
 			'odyssey_stats_changed_at' => 0,
 			'notices'                  => array(),
 			'views'                    => 0,
-			'modules'                  => array(),
 		);
 		$this->assertSame( $options_should_be, $options );
 	}
@@ -77,7 +76,6 @@ class Test_Options extends StatsBaseTestCase {
 			'odyssey_stats_changed_at' => 0,
 			'notices'                  => array(),
 			'views'                    => 0,
-			'modules'                  => array(),
 		);
 		$this->assertSame( $options_should_be, $options );
 	}
@@ -159,7 +157,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'set version'          => array(
@@ -181,7 +178,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'set blog blog_id'     => array(
@@ -203,7 +199,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'multiple options'     => array(
@@ -236,7 +231,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 		);
@@ -296,7 +290,6 @@ class Test_Options extends StatsBaseTestCase {
 			'odyssey_stats_changed_at' => 0,
 			'notices'                  => array(),
 			'views'                    => 0,
-			'modules'                  => array(),
 		);
 		$this->assertTrue( Options::set_options( $set_options ) );
 		$this->assertSame( $stored_options, get_option( 'stats_options' ) );
@@ -327,7 +320,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'Roles'        => array(
@@ -348,7 +340,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'Count Roles'  => array(
@@ -368,7 +359,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'Version'      => array(
@@ -388,7 +378,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'Honor DNT'    => array(
@@ -408,7 +397,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 			'Dummy option' => array(
@@ -428,7 +416,6 @@ class Test_Options extends StatsBaseTestCase {
 					'odyssey_stats_changed_at' => 0,
 					'notices'                  => array(),
 					'views'                    => 0,
-					'modules'                  => array(),
 				),
 			),
 		);
