+++ projects/packages/stats/tests/php/test-xmlrpc-provider.php
@@ -101,7 +101,6 @@ class Test_XMLRPC_Provider extends StatsBaseTestCase {
 			'odyssey_stats_changed_at' => 0,
 			'notices'                  => array(),
 			'views'                    => 0,
-			'modules'                  => array(),
 			'host'                     => 'example.org',
 			'path'                     => '/',
 			'blogname'                 => false,
