--- projects/plugins/jetpack/changelog/add-stats-settings-support
@@ -1,5 +1,0 @@
-Significance: patch
-Type: other
-Comment: Updated composer.lock.
-
-
