--- projects/plugins/search/changelog/add-stats-settings-support
@@ -1,5 +1,0 @@
-Significance: patch
-Type: changed
-Comment: Updated composer.lock.
-
-
